pping: Add option to aggregate RTTs
Add an option -a[X] or --aggregate[=X] to every X (X=1 be default)
provide an aggregate report of all RTT samples. This is currently
mutually exclusive with the normal per-RTT sample reports.

In this initial proof-of-concept for aggregating RTT samples instead
of reporting them one by one, the minimum and maximum RTT are tracked
and all RTTs are added to a histogram. The uses a predetermined number
of bins of equal width (by default 1000 bins, each 0.1 ms wide), see
RTT_AGG_NR_BINS and RTT_AGG_BIN_WIDTH in pping.h.

The aggregated stats are never reset, and thus contain all RTTs since
the start of tracing. The next commit will change this to reset the
stats after every report, so that each report only contain the RTTs
since the last report.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>