pping: Reset aggregated RTTs after each report
Instead of keeping all RTTs since ePPing started, reset the aggregated
stats after each time they're reported (by default once per second),
so the report only shows the RTTs since the last report.

To avoid concurrency issues due to user space reading and resetting
the map while the BPF programs are updating it, use two aggregation
maps (or rather one array-map with two indices), one active and one
inactive. Each time user space wishes to report the aggregated RTTs it
first switches which map is actively used by the BPF progs, and then
reads and resets the now inactive map.

As the RTT stats are now periodically reset, change the
histogram (aggregated_rtt_stats.bins) to now use __u32 instead of
__u64 counts as chance for overflowing is low (even if 1 million
RTTs/s is added to the same bin, it would take over an hour to
overflow, and report frequency is likely higher than that).

One noteworthy detail when it comes to the implementation is that the
map used to hold the value of which map the BPF progs should currently
aggregate the RTTs in (the active map) is __u64 (despite keys to array
lookups only being __u32, and the value only switching between 0 and
1, which would fit in a single bit). The reason for this is to make it
easier for the user space code to traverse the values, as the per-CPU
maps will 8-byte align each value (so if using ex. __u32 instead the
user space logic would have to manually 8-byte align the __u32s when
traversing the map).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>