pping: Add options to ignore RTT reports on kernel or user side
Add an option -k/--kernel-drop which causes the BPF programs to not
push any RTT report to userspace, and the option -u/--user-drop which
causes the userspace component to not print out the RTT report it
receives (only applies to standard format). This is only useful to
test overhead related to reporting the RTTs.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>